Fix shader parameter. Build cubes doesnt do color nor position correct

+++ Assets/Scripts/Generation/MarchingCubes/Pipeline/ChunkGenerationGPUData.cs
@@ -60,7 +60,7 @@ namespace MarchingCubes
             densityGeneratorShader.SetVector("anchor", anchor);
             densityGeneratorShader.SetVector("offset", offset);
             densityGeneratorShader.SetInt("numPointsPerAxis", pointsPerAxis);
-            densityGeneratorShader.SetInt("spacing", chunk.LOD);
+            densityGeneratorShader.SetFloat("spacing", chunk.LOD);
             densityGeneratorShader.SetBool("tryLoadData", tryLoadData);
         }
